fix: preserve Amp thread project paths
