+++ src/adapters/paths.rs
@@ -23,6 +23,49 @@ pub(crate) fn resolve_home_dir(
     Ok(Some(dir))
 }
 
+pub(crate) fn file_uri_to_path(uri: &str) -> Option<String> {
+    let rest = uri.strip_prefix("file://")?;
+    let decoded = percent_decode(rest)?;
+    if cfg!(windows)
+        && let Some(drive) = decoded.strip_prefix('/')
+        && drive.len() >= 2
+        && drive.as_bytes()[1] == b':'
+    {
+        return Some(drive.to_string());
+    }
+    Some(decoded)
+}
+
+fn percent_decode(input: &str) -> Option<String> {
+    let bytes = input.as_bytes();
+    let mut out = Vec::with_capacity(bytes.len());
+    let mut index = 0;
+    while index < bytes.len() {
+        if bytes[index] == b'%' {
+            if index + 2 >= bytes.len() {
+                return None;
+            }
+            let hi = from_hex(bytes[index + 1])?;
+            let lo = from_hex(bytes[index + 2])?;
+            out.push((hi << 4) | lo);
+            index += 3;
+        } else {
+            out.push(bytes[index]);
+            index += 1;
+        }
+    }
+    String::from_utf8(out).ok()
+}
+
+fn from_hex(byte: u8) -> Option<u8> {
+    match byte {
+        b'0'..=b'9' => Some(byte - b'0'),
+        b'a'..=b'f' => Some(byte - b'a' + 10),
+        b'A'..=b'F' => Some(byte - b'A' + 10),
+        _ => None,
+    }
+}
+
 pub(crate) fn vscode_extension_task_dirs(extension_id: &str) -> Vec<PathBuf> {
     vscode_extension_task_dirs_from(dirs::config_dir(), extension_id)
 }
@@ -67,6 +110,15 @@ mod tests {
     #[test]
     fn missing_config_dir_returns_empty() {
         assert!(vscode_extension_task_dirs_from(None, "saoudrizwan.claude-dev").is_empty());
+    }
+
+    #[test]
+    fn file_uri_path_decodes_percent_escapes() {
+        assert_eq!(
+            file_uri_to_path("file:///Users/x/git/foo%20bar").as_deref(),
+            Some("/Users/x/git/foo bar")
+        );
+        assert_eq!(file_uri_to_path("https://example.com"), None);
     }
 
     #[test]
